Changes for mParticle PR
