+++ build.gradle
@@ -25,6 +25,6 @@ android {
 dependencies {
     compile 'com.android.support:appcompat-v7:24.2.1'
     compile 'com.android.support:recyclerview-v7:24.2.1'
-    compile 'com.rokolabs:rokomobi-android-sdk:1.12.+@aar'
+    compile 'io.instabot:instabot-android-sdk:1.0.0@aar'
     compile 'com.google.code.gson:gson:2.7'
 }
