+++ sample/build.gradle
@@ -1,34 +1,64 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    
     repositories {
-        mavenLocal()
-        google()
         jcenter()
+        maven {
+            url 'https://maven.google.com/'
+            name 'Google'
+        }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.0-beta1'
         classpath 'com.google.gms:google-services:3.0.0'
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
     }
 }
 
-allprojects {
-    repositories {
-        mavenLocal()
-        google()
-        jcenter()
+
+apply plugin: 'com.android.application'
+
+android {
+    compileSdkVersion 25
+    buildToolsVersion "25.0.2"
+    defaultConfig {
+        applicationId "com.rokolabs.mparticle.sample"
+        minSdkVersion 23
+        targetSdkVersion 25
+        versionCode 1
+        versionName "1.0"
+        multiDexEnabled true
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    lintOptions {
+        abortOnError false
     }
 }
 
-task clean(type: Delete) {
-    delete rootProject.buildDir
+repositories {
+    maven {
+        url 'https://artifactory.roko.mobi/artifactory/libs-release'
+    }
+
+    flatDir {
+        dirs 'libs'
+    }
+    maven {
+        url 'https://maven.google.com/'
+        name 'Google'
+    }
 }
 
-configurations.all {
-    // check for updates every build
-    resolutionStrategy.cacheChangingModulesFor 0, 'seconds'
+dependencies {
+    compile 'com.mparticle:android-core:4.15.4-SNAPSHOT'
+    compile 'com.mparticle:android-instabot-kit:4.15.4-SNAPSHOT'
+    compile 'com.android.support:appcompat-v7:25.4.0'
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    compile 'com.android.support:design:25.4.0'
+    compile 'com.google.firebase:firebase-messaging:9.8.0'
+    compile 'com.android.support:multidex:1.0.2'
 }
+apply plugin: 'com.google.gms.google-services'
