+++ sample/src/main/AndroidManifest.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="io.instabot.mparticle.sample">
+
+    <application
+        android:name=".MParticleSampleApplication"
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:theme="@style/AppTheme">
+        <activity
+            android:name=".MainActivity"
+            android:label="@string/app_name"
+            android:theme="@style/AppTheme.NoActionBar">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity android:name=".EventsActivity" />
+        <activity android:name=".UserActivity" />
+        <activity android:name=".InstabotUIActivity">
+            <intent-filter>
+                <action android:name="io.instabot.INSTABOT_ACTIVITY" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity>
+
+        <receiver android:name="com.mparticle.MPReceiver"
+            android:permission="com.google.android.c2dm.permission.SEND">
+            <intent-filter>
+                <action android:name="com.google.android.c2dm.intent.RECEIVE" />
+                <!-- Use your package name as the category -->
+                <category android:name="com.mparticle.kits.pushsample" />
+            </intent-filter>
+        </receiver>
+
+        <service android:name="com.mparticle.MPService" />
+    </application>
+</manifest>
