+++ sample/src/main/java/io/instabot/mparticle/sample/MainActivity.java
@@ -1,4 +1,4 @@
-package com.rokolabs.mparticle.sample;
+package io.instabot.mparticle.sample;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,16 +7,6 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
-import android.widget.TextView;
-
-import com.mparticle.DeepLinkError;
-import com.mparticle.DeepLinkListener;
-import com.mparticle.DeepLinkResult;
-import com.mparticle.MParticle;
-import com.rokolabs.sdk.analytics.Event;
-import com.rokolabs.sdk.analytics.RokoLogger;
-import com.rokolabs.sdk.instabot.Instabot;
-import com.rokolabs.sdk.tools.RokoTools;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -26,8 +16,6 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        ((TextView) findViewById(R.id.app_version)).setText("App: v"+BuildConfig.VERSION_NAME);
-        ((TextView) findViewById(R.id.sdk_version)).setText("SDK: v"+ com.rokolabs.sdk.BuildConfig.VERSION_NAME);
 
         ListView list = (ListView) findViewById(R.id.list);
         list.setOnItemClickListener(new AdapterView.OnItemClickListener() {
