+++ sample/src/main/java/io/instabot/mparticle/sample/UserActivity.java
@@ -1,4 +1,4 @@
-package com.rokolabs.mparticle.sample;
+package io.instabot.mparticle.sample;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
