+++ sample/src/main/res/layout/activity_main.xml
@@ -3,7 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context="com.example.sobolev.mparticle.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
