+++ sample/src/main/res/layout/events_activity.xml
