+++ sample/src/main/res/layout/instabot_ui_activity.xml
