+++ sample/src/main/res/layout/user_activity.xml
