+++ sample/src/main/res/values/arrays.xml
