+++ sample/src/main/res/values/mparticle.xml
