+++ sample/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">MParticle ROKO Mobi Sample App</string>
+    <string name="app_name">MParticle Instabot Sample App</string>
     <string name="action_settings">Settings</string>
 </resources>
