+++ src/main/java/com/mparticle/kits/InstabotProvider.java
@@ -1,0 +1,5 @@
+package com.mparticle.kits;
+
+public interface InstabotProvider {
+    io.instabot.sdk.Instabot instabot();
+}
