--- src/main/java/com/mparticle/kits/RokoMobiProvider.java
@@ -1,7 +1,0 @@
-package com.mparticle.kits;
-
-import com.rokolabs.sdk.instabot.Instabot;
-
-public interface RokoMobiProvider {
-    Instabot instabot();
-}
