+++ src/test/java/com/mparticle/kits/KitTests.java
@@ -1,12 +1,8 @@
 package com.mparticle.kits;
 
 
-import android.content.Context;
-
 import org.junit.Test;
-import org.mockito.Mockito;
 
-import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.Assert.assertNotNull;
@@ -16,7 +12,7 @@ import static org.junit.Assert.fail;
 public class KitTests {
 
     private KitIntegration getKit() {
-        return new RokoMobiKit();
+        return new InstabotKit();
     }
 
     @Test
